Set up automatic Ajax passthrough for server-only Sherpa methods
This feature allows using sherpa functions only available in the main server-side script include. It works by first checking for a client-side implementations when attempting to access a library function; if there was no client-side implementation, an Ajax call is fired to the server that invokes the same sherpa function on the server-side.

When a function is accessed in this way, the return value will always be wrapped in a Promise